Split game script code
