+++ src/utils.ts
@@ -1,4 +1,6 @@
 export interface Coordinates {
-    x: number;
-    y: number;
+  x: number;
+  y: number;
 }
+
+export const UTF16_CODE_OF_LETTER_A = 65;
